Improve the condition and merge tests into existing file
